test: split cron service regression ownership
